+++ src/main.cpp
@@ -2,11 +2,15 @@
 #include <WiFi.h>
 #include <FS.h>
 #include <SD.h>
+#include <ESP32_VS1053_Stream.h>
 
-#include "websocketEventHandler.h"
+#include "playList.h"
 #include "icons.h"
 #include "percentEncode.h"
 #include "WiFiCredentials.h" /* untracked file in folder include */
+#include "tftMessage_t.h"
+#include "playerMessage_t.h"
+#include "serverMessage_t.h"
 
 const char *PROGRAM_NAME = "featherplayer-esp32";
 const char *FAVORITES_FOLDER = "/"; /* if this is a folder use a closing slash */
@@ -20,6 +24,10 @@ SemaphoreHandle_t spiMutex = nullptr; // SPI bus is shared between playertask -V
 QueueHandle_t tftQueue = nullptr;
 QueueHandle_t playerQueue = nullptr;
 QueueHandle_t serverQueue = nullptr;
+
+extern int SDREADER_CS;
+extern void serverTask(void *parameter);
+extern void tftTask(void *parameter);
 
 void mountSDcard()
 {
